+++ src/services/traits/AccessorByIdOrString.php
@@ -15,6 +15,8 @@ use yii\db\ActiveRecord as Record;
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
  * @since 1.0.0
+ *
+ * @deprecated
  */
 trait AccessorByIdOrString
 {