+++ src/services/traits/ActiveRecordByString.php
@@ -14,6 +14,8 @@ use flipbox\ember\exceptions\RecordNotFoundException;
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
  * @since 1.0.0
+ *
+ * @deprecated
  */
 trait ActiveRecordByString
 {