+++ src/services/traits/IdAccessor.php
@@ -17,6 +17,8 @@ use yii\db\ActiveRecord as Record;
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
  * @since 1.0.0
+ *
+ * @deprecated
  */
 trait IdAccessor
 {